Updated Notifications / Layout
Got the notification for friend requests to display under the user
button as well as the number of new friend requests to display within
the button for friend requests. I touched on making the layouts a
little more consistent but will need a lot more time to really do any
styling. Perhaps I will get to that this week.