Friends Scrollable and Clickable
Couldn’t test that scrollable worked because of emulator and need to
add some more friends but I believe it worked, I cleaned up how it was
done a lot. Also messed around with the icons but they need a lot of
work, for one I’m taking smiles off the faces and going to be messing
around with colors and etc.