Notifications / Worked on Scaling / Refactoring
Began implementation of a different way of doing the notifications. Put
in client side / server side code for getting number of groups /
invites. (Doesn't work yet, I will resume work tomorrow). Took a look
into scaling and changed a few things, will touch on this again tomorrow
as well. Also I did some refactoring, noticed that the getJSONFeed for
many calls was the same so I made it all accessible from global. I will
look into making all of the calls to that conform. Been messing around a
bit more with the colors, this is a work in progress and feel free for
input on anything regarding layout. There are a decent amount of things
that I'll be addressing this week with the scalability of the layout and
also be touching up the style along the way.